refactor: add SearchWrap component
